==> notebooks/01_ToxVal_processing.ipynb <==
raw_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/raw/'
processed_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/processed/'
interim_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/interim/'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import os
import glob

for f in os.listdir(raw_dat_dir):
    print(f)

def loadTXdata(file, dat_dir = raw_dat_dir):
    df = pd.read_csv(dat_dir+file)
    return  df

toxval = loadTXdata('Original_ToxVal_v7_24-04-2019.csv')
toxval.head()


toxval[toxval['name'] == 'Ethylbenzene'].shape

toxval.columns.tolist()

cols_to_drop = ['chemical_id',
 'toxval_id','toxval_uuid','parent_toxval_uuid','species_id',
 'study_id','priority_id','manual_qc_flag','source_url','subsource_url','record_url','source_source_id','toxval_numeric_converted','toxval_units_converted','chemical_id_external']

#len(cols_to_drop)

toxval.drop(cols_to_drop, axis = 1, inplace = True)

toxval.shape

mycols = toxval.columns.tolist()

mycols2 = mycols[45:] + mycols[:45]

#mycols2

toxval = toxval[mycols2]

#toxval[toxval.name == 'Ethylbenzene']

toxval.risk_assessment_class.unique()

study_length = []
for x, y, z, a in list(zip(toxval['risk_assessment_class'], toxval['study_type'], toxval['study_duration_value'], toxval['study_duration_units'])):
    if (x in ['short-term','repeat dose'] and y == 'chronic'):
        study_length.append('chronic')
    elif (x in ['short-term','repeat dose']) and (y not in ["reproductive", "subchronic"]) and (z >=100) and (a == "day"):
        study_length.append('chronic')
    elif (x in ["short-term", "repeat dose"]) and (y not in ["reproductive", "subchronic"]) and (z >=14) and (a == "week"):
        study_length.append('chronic')
    elif (x in ["short-term", "repeat dose"]) and (y not in ["reproductive", "subchronic"]) and (z>=4) and (a == "month"):
        study_length.append('chronic')
    elif (x in ["short-term", "repeat dose"]) and (y not in ["reproductive", "subchronic"]) and (z >=1) and (a == "year"):
        study_length.append('chronic')
    elif (x in ["short-term", "repeat dose"]) and  (y == "subchronic"):
        study_length.append('subchronic')
    elif (x in ['short-term','repeat dose']) and (y not in ["reproductive", "chronic"]) and (z>=30) and  (z <100) and (a == "day"):
        study_length.append('subchronic')
    elif (x in ['short-term','repeat dose']) and (y not in ["reproductive", "chronic"]) and (z>=5) and  (z <14) and (a == "week"):
        study_length.append('subchronic')
    elif (x in ['short-term','repeat dose']) and (y not in ["reproductive", "chronic"]) and (z>=1.5) and  (z <4) and (a == "month"):
        study_length.append('subchronic')
    elif (x == 'repeat dose') and (y not in ["acute", "reproductive", "chronic"]) and (z==-1):
        study_length.append('subchronic')
    elif (x == 'repeat dose') and (y not in ["acute", "reproductive", "chronic"]) and (z==0):
        study_length.append('subchronic')
    elif (x in ['short-term','repeat dose']) and (y == "reproductive"):
        study_length.append('reproductive')
    else:
        study_length.append(x)
        

toxval['study_length'] = study_length

len(study_length)

#toxval.to_csv(processed_dat_dir+'ToxVal_wstudylength.csv')

toxval.head()

toxval.dsstox_substance_id.nunique()

toxval.qsar_ready_smiles.nunique()
