==> notebooks/03_chemprop_alerts.ipynb <==
raw_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/raw/'
processed_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/processed/'
interim_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/interim/'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import os
import glob

### Import chemical properties

chem_prop = pd.read_csv(processed_dat_dir+'ToxVal_v7_chem_prop.csv')

chem_prop.shape

toxval = pd.read_csv(processed_dat_dir+'ToxVal_wstudylength.csv')

#toxval.columns.tolist()
#toxval.toxval_units.unique()

#toxval.exposure_route.unique()

toxval_chemprop = pd.merge(toxval, chem_prop, on = 'DTXSID', how = 'left')

toxval_chemprop.shape

### ppm to mgm3 unit conversions: Basic equation for conversions: mg/m3 = ppm* Mol Weight/24.45

mgm3_units = []
for x, y, z, a in list(zip(toxval_chemprop['exposure_route'], toxval_chemprop['toxval_units'], toxval_chemprop['MolWeight'], toxval_chemprop['toxval_numeric'])):
    if (x == 'inhalation' and y == 'ppm') or (x == 'inhalation' and y == 'mg/L'):
        xx = (a*z)/24.45
        mgm3_units.append(xx)
    elif (x == "inhalation" and y == "g/m3"):
        mgm3_units.append(a * 1000)
    elif (x == "inhalation" and y == "ug/m3"):
        mgm3_units.append(a / 1000)
    elif (x == "inhalation" and y == "mg/m") or (x == "inhalation" and y == "mg/m3"):
        mgm3_units.append(a)
    else:
        mgm3_units.append(a)
    

len(mgm3_units)

### mg/m3 to ppm conversion: Basic equation used: ppm = (24.45 * mg/m3) / Mol Weight

ppm_units = []
for x, y, z, a in list(zip(toxval_chemprop['exposure_route'], toxval_chemprop['toxval_units'], toxval_chemprop['MolWeight'], toxval_chemprop['toxval_numeric'])):
    if (x == 'inhalation' and y == 'mg/m') or (x == 'inhalation' and y == 'mg/m3'):
        xx = (a*24.45)/z
        ppm_units.append(xx)
    elif (x == "inhalation" and y == "g/m3"):
        xy = (24.45 * (a * 1000)) / z
        ppm_units.append(xy)
    elif (x == "inhalation" and y == "ug/m3"):
        bb = (24.45 * (a/1000)) / z
        ppm_units.append(bb)
    elif (x == "inhalation" and y == "ppm") or (x == "inhalation" and y == "mg/L"):
        ppm_units.append(a)
    else:
        ppm_units.append(a)
    

len(ppm_units)

toxval_chemprop['toxval_ppm'] = ppm_units
toxval_chemprop['toxval_mgm3'] = mgm3_units

toolbox = pd.read_csv(processed_dat_dir+'Toolbox_out.csv')

toolbox.head()

import re
p = re.compile(r'DTXSID\d{1,}')

dtxsid = [m.group(0) for l in toolbox['Chemical name(s)'] for m in [p.search(l)] if m]

len(dtxsid)

toolbox = toolbox[toolbox['Chemical name(s)'].str.contains('DTXSID')]

toolbox['DTXSID'] = dtxsid

toolbox.drop(['CAS', 'Chemical name(s)', 'Additional Ids', 'Composition', 'CAS Smiles relation'], axis = 1, inplace = True)


toolbox = toolbox[['DTXSID','SMILES','Eye irritation/corrosion Exclusion rules by BfR',
       'Eye irritation/corrosion Inclusion rules by BfR',
       'Respiratory sensitisation',
       'Skin irritation/corrosion Exclusion rules by BfR',
       'Skin irritation/corrosion Inclusion rules by BfR' ]]

toolbox.columns = ['DTXSID','SMILES','BfR_eye_exclusion', 'BfR_eye_inclusion', 'resp_sens', 'BfR_skin_exclusion', 'BfR_skin_inclusion']

toolbox.head()

toolbox['BfR_skin_inclusion'] = [e.replace(';','|') for e in toolbox['BfR_skin_inclusion'] ]


toolbox['BfR_eye_inclusion'] = [e.replace(';','|') for e in toolbox['BfR_eye_inclusion'] ]

toxval_chemprop.shape

toxval_chemprop[toxval_chemprop['DTXSID'] == 'DTXSID3020596'].shape

toxval_prop = pd.merge(toxval_chemprop, toolbox, on = 'DTXSID', how = 'left')

toxval_prop.to_csv(processed_dat_dir+'ToxVal_wpred_properties.csv')

toxval_prop.shape

toxval_prop.head()

toxval_prop  = pd.read_csv(processed_dat_dir+'ToxVal_wpred_properties.csv')

toxval_prop.shape
