==> notebooks/04_subsetting_for_inhalation.ipynb <==
raw_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/raw/'
processed_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/processed/'
interim_dat_dir = '/home/grace/Documents/python/ttc_inhalation/data/interim/'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import os
import glob

### Import Cramer classes

cramer = pd.read_csv(interim_dat_dir+'Cramer_4Toxtree_results.csv')
cramer.head()

### Import Toxval data

toxval_prop = pd.read_csv(processed_dat_dir+'ToxVal_wpred_properties.csv')

toxval_prop[toxval_prop['DTXSID'] == 'DTXSID3020596'].shape


toxval_prop.describe()

def filter_ambigous(df):
    mask = (df['toxval_numeric'] == 0) | (df['toxval_numeric'] == -999)
    return df[~mask]

toxval_prop = filter_ambigous(toxval_prop)

toxval_prop.describe()

toxval_prop['study_length'].unique()

toxval_prop.columns.tolist()
toxval_prop = toxval_prop[[
 'casrn',
 'name',
 'DTXSID',
 'dsstox_compound_id',
 'qsar_ready_smiles',
 'toxval_ppm',
 'toxval_mgm3',
 'source',
 'subsource',
 'toxval_type',
 'toxval_subtype',
 'toxval_numeric',
 'toxval_units',
 'risk_assessment_class',
 'study_type',
 'study_duration_class',
 'study_duration_value',
 'study_duration_units',
 'species_common',
 'strain',
 'sex',
 'sex_original',
 'population',
 'exposure_route',
 'exposure_method',
 'exposure_form',
 'media',
 'year',
 'human_eco',
 'critical_effect',
 'lifestage',
 'generation',
 'study_length',
 'MolWeight',
 'nbAtoms',
 'nbHeavyAtoms',
 'nbC',
 'nbO',
 'nbN',
 'nbAromAtom',
 'nbRing',
 'nbHeteroRing',
 'Sp3Sp2HybRatio',
 'nbRotBd',
 'nbHBdAcc',
 'ndHBdDon',
 'nbLipinskiFailures',
 'TopoPolSurfAir',
 'MolarRefract',
 'CombDipolPolariz',
 'logP',
 'AD_LogP',
 'AD_index_LogP',
 'Conf_index_LogP',
 'melting_point',
 'AD_MP',
 'AD_index_MP',
 'Conf_index_MP',
 'boiling_point',
 'AD_BP',
 'AD_index_BP',
 'Conf_index_BP',
 'log_vapour_pres',
 'AD_VP',
 'AD_index_VP',
 'Conf_index_VP',
 'log_water_sol',
 'AD_WS',
 'AD_index_WS',
 'Conf_index_WS',
 'log_henrys_law',
 'AD_HL',
 'AD_index_HL',
 'Conf_index_HL',
 'logKOA',
 'AD_KOA',
 'AD_index_KOA',
 'Conf_index_KOA',
 'ionization',
 'pKa_acid',
 'pKa_base',
 'AD_pKa',
 'AD_index_pKa',
 'Conf_index_pKa',
 'logD_55',
 'logD_74',
 'AD_LogD',
 'AD_index_LogD',
 'Conf_index_LogD',
 'logOH',
 'AD_AOH',
 'AD_index_AOH',
 'Conf_index_AOH',
 
 'SMILES',
 'BfR_eye_exclusion',
 'BfR_eye_inclusion',
 'resp_sens',
 'BfR_skin_exclusion',
 'BfR_skin_inclusion']]

adj_unit = []
for x, y in list(zip(toxval_prop['study_length'],toxval_prop['toxval_ppm'] )):
    if x == 'subchronic':
        adj_unit.append(y/2)
    elif x == 'subacute':
        adj_unit.append(y/6)
    else:
        adj_unit.append(y)
toxval_prop['toxval_ppm_adj'] = adj_unit

adj_units = []
for x, y in list(zip(toxval_prop['study_length'],toxval_prop['toxval_mgm3'] )):
    if x == 'subchronic':
        adj_units.append(y/2)
    elif x == 'subacute':
        adj_units.append(y/6)
    else:
        adj_units.append(y)
toxval_prop['toxval_mgm3_adj'] = adj_units

toxval_prop.shape

cramer['Cramer rules'].value_counts()

ClassI = cramer[cramer['Cramer rules'] == 'Low (Class I)']['DTXSID'].tolist()
ClassII = cramer[cramer['Cramer rules'] == 'Intermediate (Class II)']['DTXSID'].tolist()
ClassIII = cramer[cramer['Cramer rules'] == 'High (Class III)']['DTXSID'].tolist()


#toxval_prop['species_common']

import re

#regex = re.compile('\s{1,}rat\s{1,}')
#set([m.group(0) for l in toxval_prop['species_common'] for m in [regex.search(l)] if m]) 

#regex = re.compile('other:\s{1,}')
#[m.group(0) for l in toxval_prop['species_common'] for m in [regex.search(l)] if m] 
#toxval_prop['species'] = toxval_prop['species_common'].str.rstrip('other:')

toxval_prop['species_common'] = toxval_prop['species_common'].str.replace(' -','')

toxval_prop['species_common'] = toxval_prop['species_common'].str.replace('other: ', '')

rodent_dict = dict(rodent = ['mouse, rat',
 'rat','rat & dog',
 'white rats',
 'rats','albino rat', 'albino laboratory rat','mouse, rat, guinea pig','rat, young',
 'rat, dog',
 'mouse / rat',
 'rat / mouse',
 'mouse , albino rat , guinea pig  and rabbit',
 'rat, rabbit',
 'rat and mouse',
 'rattus norvegicus','rattus rattus','rattus argentiventer','rattus exulans','rats (female)',
 'rabbits and rats',
 'rats and mice',
 'rats (male)',
 'rats (wistar strain for both studies)',
 'rat, rabbit and mouse',
 'mice, rats, and rabbits',
 'rat and guinea pig','mouse and rat',
 'rat, mouse, rabbit, human',
 'rabbit, rat, and cat',
 'cats, rabbits, guinea pigs, rats, mice',
 'rat and rabbit', 'rat, mouse, guinea pig',
 'rat and rabbits',
 'rats and guinea pigs',
 'rabbits, rats, guinea pigs, cats, dogs and monkeys.',
 'rat, rabbit, guinea pig',
 'not reported (rats, mice, guinea pigs and/or rabbits)',
 '646 white rats, 57 guinea pigs, 102 white mice and 12 rabbits.','rattus tiomanicus','cat, dog, guinea pig, rabbit, rat    ','fischer 344 rats, b6c3f1 mice and yellow and black c5ys f1 mice.',
 'rabbit and rat',
 ' rats and guinea pigs',
 'rats, hamsters, guinea pigs',
 'rats and dogs (only females)','rat & mouse',
 'rat and mice',
 'rat, rabbit, dog',
 'rat / mice','rat, mouse, monkey, guinea-pig, dog','mice & rats',
 '1 rabbit, 1 guinea pig, 2 rats and 4 mice','rat, mouse',
 'review of rat and mouse data',
 'rats and quinea pigs',
 'mice and rats',
 'oecd evaluation of available rat and mouse studies',
 'rat, mouse and guinea pig',
 'rat;human',
 '1 cat, 1 rabbit, 1 guinea pig, 2 rats and 4 mice','10 rats and 20 mice',
 'rat, trout',
 'mouse, rat, rabbit','rats, guinea pigs, rabbits, mice and monkeys',
 'rat and dog',
 'cat, rat, rabbit',
 'rats, guinea pigs and rabbits',
 'rat,mice',
 'rats and guinea pigs tested.',
 'rats, rabbits, guinnea pigs, monkeys.',
 'rat, guinea pig, rabbit   ',
 'rat, guinea pig, rabbit, dog','rats, guinea pigs, rabbits and dogs',
 'rats, rabbits and cats',
 'female mice, male and female rats, male guinea pigs, male rabbit, cats',
 'studies in various species (guinea pig, rat and rabbit) are summarised',
 'rats, guinea pigs, rabbits, cats, dogs and monkeys.',
 'rat, dog, mouse',
 'rats, rabbits, guinea pigs, dogs, cats and monkeys',
 'rats, guinea pigs, rabbits, mice',
 'rats, mice, guinea pig, and rabbits',
 'mice, rats, rabbits and dogs',
 'three animal species used: rat, mouse, syrian hamster',
 'rats (wistar) and guinea pig',
 'mouse, rat, dog and rabbit',
 'rats, mice, guinea pigs and rabbits',
 'rat, rabbit, cat and guinea pig',
 'rabbit, guinea pig, rat, cat',
 'dogs, guinea pigs, and rats',
 'rats, mice, hamsters, non-human primates, dogs ',
 'rabbit, guinea pig, rat',
 'rat (albino)',
 'rat, mice',
 'rat (males) and mouse (females)', 'rats, mice',
 'rats, dogs',
 'rat, guinea pig',
 'rabbit (gross e); rat; dog',
 'mouse, rat, rabbit, guinea pig',
 'laboratory animal',
 'rat, albino','white mice, albino rats, new zealand rabbits',
 'dog, rabbit, rat','cat, dog, guinea pig, rabbit, rat',
 'rat: sprague-dawley and dog: beagle ',
 'cat, rabbit, guinea pig, rat',
 'rat, rabbit, guinea pig, mouse','rats, guinea pig, dogs',
 'rats, guinea pigs, dogs',
 'rat, wistar crl:(wi) br (outbred, spf-quality)',
 'rat, wistar strain crl:(wi) br (outbred, spf quality)',
 'rat, wistar strain crl:(wi) br (outbred, spf-quality).',
 'rats (f344) and mice (b6c3f1), 5/sex/group','rats, mice, guinea pigs, rabbits, cats and dogs',
 'cat, dog, guinea pig, mouse, rabbit and rat',
 'albino rat, guinea pig, rabbit',
 'rat (female and male), mice (female and male) and monkey (male)',
 'rats (males and females) and mice (females)',
 'hamster, rat, gerbil',
 'rats, guinea pigs, rabbits, monkeys, and dogs',
 'rats and rabbits',
 'rat and albino rabbit','rabbits,rats, and guinea pigs',
 'rat and syrian hamster','mice, rats, rabbits, guinea pigs and monkeys',
 'sprague-dawley rats or new zealand white rabbits',
 'mice, rats, rabbits, cats and dogs','mice, rats and hamsters',
 'rat, hanrcc:wist (spf)',
 'rats and miniatur swine','rat, mouse, hamster',
 'rats and hamsters',
 'not specified, bodyweights corresponding to rats',
 'study 1: mice; study 2 & 3: rats',
 'study 1, study 2, study 4 & study 5: rat; study 3: mouse',
 'study 1: rat; study 2: mouse; study 3: mouse',
 'study 1: mouse; study 2: rat',
 'mouse & rat',
 'humans and rats',
 'mouse, rat and guinea-pigs','f344/n rats and b6c3f1/n mice','rat; mus musculus; mouse',
 'rattus norvegicus; mus musculus',
 'rats, mice and guinea pigs',
 'rat, guinea pig, dog','rats and humans',
 'mouse; rat',
 'in all analysed tests: rats',
 'rabbits, guinea pigs, rats and mice','rat, guinea pig, dog, cat',
 'rat, guinea pig, dog and cat',
 'rats, guinea pigs, rabbits, and  monkeys',
 'white rats and guinea pigs',
 'no definite data, probably rats',
 'mouse, rat, and guinea pig',
 'sprague-dawley rats, hartley guinea pigs, squirrel monkeys, beagle dogs',
 'albino rats, mice and guinea pigs',
 'mice, rats, guinea pigs',
 'rats & mice',
 'rat & mice',
 'rabbit, guinea pig, rats, mice',
 'rat, mouse, hamster, dog',
 'rat and hamster',
 'white mice, albino rats,  guinea pigs, rabbits','rats  and mice',
 'likely rat',
 'rats, guinea pigs, mice',
 'rat and guinea pigs',
 'rat, sprague-dawley','rat, monkey',
 'rat, mouse, rabbit, cat, dog, guinea pig.','rat/mouse',
 'rats, mice, hamsters, rabbits and guinea-pigs',
 'albino rats,  guinea pigs and rhesus monkeys',
 'rat, guinea pig, rabbit, squirrel monkey, dog',
 'rat, guinea pig, rabbit, dog, rhesus monkey',
 'rat, mouse, dog, monkey',
 'albino rats, white mice, guinea pigs, rabbits',
 'rat, mouse, hamster, rabbit','mouse, rat and guinea pig were tested',
 'swiss albino mice, wistar rats, english short-hair guinea pigs','rats abd mice',
 'mice and rats ','rat / guinea pig','rat,  guinea pig, sheep, goat',
 'mouse, guinea pig, rabbit, rat',
 'rat;mouse',
 'albino rats, guinea pigs, rabbits',
 'rats or guinea pigs',
 'rats an dguinea pigs',
 '120 rats and 110 mice',
 'rat, mouse, guinea pig, rabbit',
 'mouse, rat and guine-pigs','mice, rat, guinea pig and dog',
 'mica and rats',
 'rats, mice, rabbits',
 'rats, mice, rabbits, guinea pigs',
 'rat, mice, rabbits, dogs',
 'rat, mice, dogs',
 'rat, mouse and hamster',
 'rabbits, rats, mice, chicken embryos',
 'rat, mouse, golden hamster, gerbil, guinea-pig','mus musculus',
 'mus musculus ssp. albinus',
 'rat; mus musculus; mouse',
 'rattus norvegicus; mus musculus',
 'mus musculus (house mouse)', 'mouse', 'rabbit','rabbits'
])

dict_rodent =  {value: key for key in rodent_dict for value in rodent_dict[key]}

#dict_rodent

toxval_prop['species'] = toxval_prop['species_common'].replace(dict_rodent)

#[e for e in toxval_prop.species_common.unique().tolist() if 'rabbit' in e]
#set(toxval_prop.species.tolist())



                    

#toxval_prop['species'].value_counts()

ClassI_data = toxval_prop[toxval_prop['DTXSID'].isin(ClassI)]
ClassII_data = toxval_prop[toxval_prop['DTXSID'].isin(ClassII)]
ClassIII_data = toxval_prop[toxval_prop['DTXSID'].isin(ClassIII)]


ClassI_data.shape

#toxval_prop.species_common.value_counts()

def filter_data(df):
    
    mask = (df['study_length'].isin(['subacute','subchronic', 'repeat dose','chronic', 'reproductive developmental', 'reproductive', 'developmental', 'multigenerational reproductive'])) & (df['exposure_route'] == 'inhalation') & (df['toxval_type'].isin(['NOAEL', 'NOEL', 'NOAEC', 'NOEC'])) & (df['species'] == 'rodent') 
    
    return df[mask]

print(filter_data(ClassIII_data).shape)
print(filter_data(ClassII_data).shape)
print(filter_data(ClassI_data).shape)

ClassI_data = filter_data(ClassI_data)
ClassII_data = filter_data(ClassII_data)
ClassIII_data = filter_data(ClassIII_data)

ClassI_data.columns.tolist()[:15]

#ClassI_data[ClassI_data.toxval_ppm == -999]


def representative(s):
    if len(s) ==1:
        return float(s)
    while len(s) >1:
        q75 = pd.Series(s.quantile(0.75))
        q25 = pd.Series(s.quantile(0.25))
        iqr = q75-q25
        pt_lower = q25 - (iqr*1.5)
        pt_higher = q75 + (iqr*1.5)
         
        newset = pd.Series([e for e in s if not (e < pt_lower).any() if not (e > pt_higher).any()])
        
        #q25out = newset.quantile(0.25)
        
        #front = pd.Series([e for e in newset if e< q25out])
        
        return newset.min()


#ClassI_data[ClassI_data['DTXSID'].isin(['DTXSID1024097', 'DTXSID1024621', 'DTXSID1024912'])]


ClassI_rep_ppm = ClassI_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_ppm_adj'])).reset_index()
ClassI_rep_ppm.columns = ['DTXSID', 'min_toxval_ppm_adj']
ClassI_rep_mgm3 = ClassI_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_mgm3_adj'])).reset_index()
ClassI_rep_mgm3.columns = ['DTXSID', 'min_toxval_mgm3_adj']

ClassI_rep_ppm.to_csv(processed_dat_dir+'ClassI_rep_ppm.csv')
ClassI_rep_mgm3.to_csv(processed_dat_dir+'ClassI_rep_mgm3.csv')

ClassII_rep_ppm = ClassII_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_ppm_adj'])).reset_index()
ClassII_rep_ppm.columns = ['DTXSID', 'min_toxval_ppm_adj']

ClassII_rep_mgm3 = ClassII_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_mgm3_adj'])).reset_index()
ClassII_rep_mgm3.columns = ['DTXSID', 'min_toxval_mgm3_adj']

ClassII_rep_ppm.to_csv(processed_dat_dir+'ClassII_rep_ppm.csv')
ClassII_rep_mgm3.to_csv(processed_dat_dir+'ClassII_rep_mgm3.csv')



ClassIII_rep_ppm = ClassIII_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_ppm_adj'])).reset_index()
ClassIII_rep_ppm.columns = ['DTXSID', 'min_toxval_ppm_adj']

ClassIII_rep_mgm3 = ClassIII_data.groupby('DTXSID').apply(lambda x: representative(x['toxval_mgm3_adj'])).reset_index()
ClassIII_rep_mgm3.columns = ['DTXSID', 'min_toxval_mgm3_adj']

ClassIII_rep_ppm.to_csv(processed_dat_dir+'ClassIII_rep_ppm.csv')
ClassIII_rep_mgm3.to_csv(processed_dat_dir+'ClassIII_rep_mgm3.csv')
